refactor: Split Transaction Logic into it's own composable
