+++ src/composables/index.ts
@@ -4,6 +4,7 @@ import useNativeToken from './useNativeToken'
 import useRadix from './useRadix'
 import useStaking from './useStaking'
 import useSidebar from './useSidebar'
+import useTransactions from './useTransactions'
 import useTokenBalances from './useTokenBalances'
 import useWallet from './useWallet'
 
@@ -14,6 +15,7 @@ export {
   useRadix,
   useStaking,
   useSidebar,
+  useTransactions,
   useTokenBalances,
   useWallet
 }
