+++ src/views/Wallet/WalletTransaction.vue
@@ -15,12 +15,12 @@
         </div>
       </div>
 
-      <div v-if="!tokenBalances || !tokenBalances.tokenBalances" class="p-4 flex items-center justify-center">
+      <div v-if="!hasTokenBalances" class="p-4 flex items-center justify-center">
         {{ $t('transaction.insufficientFunds') }}
       </div>
       <form
         @submit.prevent="handleSubmit"
-        v-else-if="tokenBalances.tokenBalances.length > 0 && nativeToken && selectedCurrency"
+        v-else-if="hasTokenBalances && nativeToken && selectedCurrency"
         class="flex flex-col items-end"
       >
         <div class="bg-white rounded-md border border-rGray text-rBlack mb-8 w-full">
@@ -130,7 +130,7 @@ import FormField from '@/components/FormField.vue'
 import ButtonSubmit from '@/components/ButtonSubmit.vue'
 import LoadingIcon from '@/components/LoadingIcon.vue'
 import FormCheckbox from '@/components/FormCheckbox.vue'
-import { useNativeToken, useRadix, useTokenBalances, useWallet } from '@/composables'
+import { useNativeToken, useRadix, useTransactions, useTokenBalances, useWallet } from '@/composables'
 import { useRouter, onBeforeRouteLeave } from 'vue-router'
 import { useI18n } from 'vue-i18n'
 
@@ -155,17 +155,25 @@ const WalletTransaction = defineComponent({
     const { errors, values, meta, setErrors, resetForm } = useForm<TransactionForm>()
     const router = useRouter()
     const { radix, networkPreamble } = useRadix()
-    const { activeAddress, transactionErrorMessage, transferTokens } = useWallet(radix, router)
+    const { activeAddress, activeAccount, hardwareAccount, hardwareAccountFailedToSign } = useWallet(radix, router)
+
+    const { transactionErrorMessage, transferTokens, transactionUnsub } = useTransactions(radix, router, activeAccount.value, hardwareAccount.value, {
+      ledgerSigningError: () => {
+        hardwareAccountFailedToSign()
+      }
+    })
+
     const { t } = useI18n({ useScope: 'global' })
     const { nativeToken, nativeTokenUnsub } = useNativeToken(radix)
     const { tokenBalances, tokenBalanceFor, tokenBalancesUnsub } = useTokenBalances(radix)
 
+    console.log(tokenBalances.value?.tokenBalances)
+
     onBeforeRouteLeave(() => {
       nativeTokenUnsub()
       tokenBalancesUnsub()
+      transactionUnsub()
     })
-
-    console.log('empty??', activeAddress.value, tokenBalances.value)
 
     // Clear form input and validation errors when switching accounts
     watch(activeAddress, () => {
@@ -202,6 +210,12 @@ const WalletTransaction = defineComponent({
 
     if (nativeToken.value) setXRDByDefault(nativeToken.value)
 
+    const hasTokenBalances = computed(() => {
+      if (!tokenBalances.value?.tokenBalances) return false
+      console.log(tokenBalances.value?.tokenBalances.length > 0)
+      return tokenBalances.value?.tokenBalances.length > 0
+    })
+
     const selectedCurrency: ComputedRef<TokenBalance | null> = computed(() => {
       if (!tokenBalances.value || tokenBalances.value.tokenBalances.length <= 0) return null
 
@@ -228,6 +242,7 @@ const WalletTransaction = defineComponent({
       errors,
       values,
       meta,
+      hasTokenBalances,
       setErrors,
       currency,
       tokenOptions,
